+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -6,11 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [12.1.2] - 2021-10-22
 
-Version Updated
-The version number for this package has increased due to a version update of a related graphics package.
+### Fixed
+- Fixed issue with mixed cached shadow for area lights.
 
 ## [12.1.1] - 2021-10-04
 
+### Fixed
 - Fixed decal position when created from context menu. (case 1368987)
 - Fixed the clouds not taking properly into account the fog when in distant mode and with a close far plane (case 1367993).
 - Fixed selection of light types (point, area, directional) for path-traced Unlit shadow mattes.